test_vblk: Fixed sign-compare

+++ tests/test_vblk.c
@@ -31,7 +31,7 @@ size_t compare_buffers(char *expected, char *actual, size_t nbytes)
 {
 	size_t diff = 0;
 
-	for (int i = 0; i < nbytes; ++i) {
+	for (size_t i = 0; i < nbytes; ++i) {
 		if (expected[i] != actual[i]) {
 			++diff;
 		}
@@ -43,9 +43,9 @@ size_t compare_buffers(char *expected, char *actual, size_t nbytes)
 void print_mismatch(char *expected, char *actual, size_t nbytes)
 {
 	printf("MISMATCHES:\n");
-	for (int i = 0; i < nbytes; ++i) {
+	for (size_t i = 0; i < nbytes; ++i) {
 		if (expected[i] != actual[i]) {
-			printf("i(%06d), expected(%c) != actual(%02d|0x%02x|%c)\n",
+			printf("i(%06lu), expected(%c) != actual(%02d|0x%02x|%c)\n",
 				i, expected[i], (int)actual[i], (int)actual[i], actual[i]);
 		}
 	}
@@ -199,7 +199,7 @@ void test_VBLK_RAND(void)
 				free(buf);
 				return;
 			}
-			CU_ASSERT(res == count);
+			CU_ASSERT(res == (ssize_t)count);
 
 			if (compare_buffers(buf_w + offset, buf, count)) {
 				if (VERBOSE)
